+++ js/card.js
@@ -1,4 +1,4 @@
-import { similarPromo } from './data.js';
+import {} from './data.js';
 
 // console.log(similarPromo);
 
@@ -17,8 +17,50 @@ const getOfferType = (promoType) => {
       return 'Дворец';
   }
 }
-let fragment = document.createDocumentFragment();
-similarPromo.forEach((promo) => {
+
+// let fragment = document.createDocumentFragment();
+// similarPromo.forEach((promo) => {
+//   const promoElement = similarPromoTemplate.cloneNode(true);
+//   promoElement.querySelector('.popup__title').textContent = promo.offer.title;
+//   promoElement.querySelector('.popup__text--address').textContent = promo.offer.address;
+//   promoElement.querySelector('.popup__text--price').textContent = promo.offer.price + ' ₽/ночь';
+
+//   promoElement.querySelector('.popup__type').textContent = getOfferType(promo.offer.type);
+
+//   promoElement.querySelector('.popup__text--capacity').textContent =
+//     `${promo.offer.rooms} комнаты для ${promo.offer.guests} гостей`;
+
+//   promoElement.querySelector('.popup__text--time').textContent =
+//     `Заезд после ${promo.offer.checkin}, выезд до ${promo.offer.checkout}`;
+
+//   let ul = promoElement.querySelector('.popup__features');
+//   ul.innerHTML = '';
+//   for (let i = 0; i < promo.offer.features.length; i++) {
+//     let li = document.createElement('li');
+//     li.classList.add('popup__feature');
+//     li.classList.add('popup__feature--' + promo.offer.features[i]);
+//     li.textContent = promo.offer.features[i];
+//     ul.appendChild(li);
+//   }
+
+//   let divPhotos = promoElement.querySelector('.popup__photos');
+//   let popupPhoto = promoElement.querySelector('.popup__photo');
+//   divPhotos.innerHTML = '';
+//   for (let i = 0; i < promo.offer.photos.length; i++) {
+//     popupPhoto.src = promo.offer.photos[i];
+//     divPhotos.appendChild(popupPhoto.cloneNode(true))
+//   }
+
+//   promoElement.querySelector('.popup__description').textContent = promo.offer.description;
+
+
+//   promoElement.querySelector('.popup__avatar').src = promo.author.avatar
+//   // similarListElement.appendChild(promoElement);
+//   fragment.appendChild(promoElement)
+// });
+
+
+const createCustomPopup = (promo) => {
   const promoElement = similarPromoTemplate.cloneNode(true);
   promoElement.querySelector('.popup__title').textContent = promo.offer.title;
   promoElement.querySelector('.popup__text--address').textContent = promo.offer.address;
@@ -41,6 +83,9 @@ similarPromo.forEach((promo) => {
     li.textContent = promo.offer.features[i];
     ul.appendChild(li);
   }
+  if (ul.children.length === 0) {
+    ul.style.display = 'none'
+  }
 
   let divPhotos = promoElement.querySelector('.popup__photos');
   let popupPhoto = promoElement.querySelector('.popup__photo');
@@ -53,12 +98,12 @@ similarPromo.forEach((promo) => {
   promoElement.querySelector('.popup__description').textContent = promo.offer.description;
 
 
-  promoElement.querySelector('.popup__avatar').src = promo.author.avatar
-  // similarListElement.appendChild(promoElement);
-  fragment.appendChild(promoElement)
-});
+  promoElement.querySelector('.popup__avatar').src = promo.author.avatar;
+
+  return promoElement;
+}
 
 // similarListElement.appendChild(fragment.children[1])
 
 // console.log(similarPromo);
-export {similarPromo, similarPromoTemplate, getOfferType}
+export {createCustomPopup}
