+++ js/main.js
@@ -1,4 +1,4 @@
 import './data.js';
 import './card.js';
 import './form.js';
-import './leaflet.js';
+import './map.js';
